requirements.txt
# Core framework
fastapi==0.115.0
uvicorn==0.30.6

# Environment & utilities
python-dotenv==1.0.1
httpx==0.27.2

# Database
pymongo==4.8.0

# Authentication & password hashing
passlib[bcrypt]==1.7.4
pyjwt==2.9.0

# OpenAI API client
openai==1.47.0

# Optional: async MongoDB (if you switch from pymongo)
# motor==3.6.0

# Type hints and extras
pydantic==2.9.2
pydantic-settings==2.4.0

# Date/time handling
python-dateutil==2.9.0.post0
